Add restrictions to parts of app for different users

+++ base/views.py
@@ -1,5 +1,7 @@
+from django.http.response import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -8,6 +10,9 @@ from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+        
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -57,6 +62,7 @@ def room(request, id):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url = 'login')
 def createRoom(request):
     form = RoomForm()
 
@@ -69,9 +75,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def updateRoom(request, id):
     room = Room.objects.get(id = id)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed HERREEEE!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -83,10 +93,12 @@ def updateRoom(request, id):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url = 'login')
 def deleteRoom(request, id):
     room = Room.objects.get(id = id)
-    
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed HERREEEE!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
